requirements.txt
# Core dependencies
python-dotenv==1.0.0
pydantic==2.5.3
aiohttp==3.12.15
PyYAML==6.0.2
APScheduler==3.11.0

# Database (mock Supabase)
sqlalchemy==2.0.25
alembic==1.13.1

# API clients
google-api-python-client==2.111.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
requests==2.31.0
httpx==0.26.0

# Scheduling and async
# apscheduler is already included above

# Logging and monitoring
structlog==23.2.0
colorlog==6.8.0
psutil==5.9.8

# Configuration
# pyyaml is already included above
toml==0.10.2

# Testing
pytest==7.4.3
pytest-asyncio==0.23.2
pytest-mock==3.12.0
httpx-mock==0.7.0

# Development
black==23.12.1
flake8==7.0.0
mypy==1.8.0
pre-commit==3.6.0